requirements.txt
# AI MINDS Backend Dependencies
flask>=3.0
flask-cors>=4.0
requests>=2.31
chromadb>=0.5
sentence-transformers>=3.0
numpy>=1.24
scipy>=1.11
rank-bm25>=0.2.2
Pillow>=10.0
pdfminer.six>=20221105
python-docx>=1.1
pyperclip>=1.8
streamlit>=1.35
